# openapi.yaml
openapi: 3.1.0
servers:
  - url: https://backend.dm.teamofsilicons.com
info: {title: DM API, version: 0.10.0}
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sync:
    get:
      operationId: syncEvents
      tags: [Delivery]
      summary: Recover authoritative event references for the authenticated actor.
      description: >-
        Returns bounded references under current conversation and group permissions. Cursors
        bind the actor, organization, testing environment and cleaning generation, and expire
        after 24 hours. Continue with the returned cursor while has_more is true; filtered
        inaccessible events can produce an empty page with more work. A Ting event sequence
        must never replace this cursor. A 409 sync_reset_required means obtain a reset cursor,
        refresh authorized conversations and messages, then resume from that cursor.
      parameters:
        - {$ref: '#/components/parameters/OrgId'}
        - {$ref: '#/components/parameters/TestingKey'}
        - {$ref: '#/components/parameters/TestingGeneration'}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
        - name: reset
          in: query
          schema: {type: boolean, default: false}
          description: Return an empty event page with a cursor at the current head. Cannot be combined with cursor.
      responses:
        '200':
          description: Current authorized references and the next authoritative cursor.
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {type: string, const: sync}
                  data: {$ref: '#/components/schemas/SyncPage'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/presence/devices/{device_id}:
    parameters:
      - {$ref: '#/components/parameters/OrgId'}
      - {$ref: '#/components/parameters/TestingKey'}
      - {$ref: '#/components/parameters/TestingGeneration'}
      - name: device_id
        in: path
        required: true
        schema: {$ref: '#/components/schemas/DeviceId'}
        description: Stable device identifier without control characters, scoped to the authenticated actor.
    put:
      operationId: renewDevicePresence
      tags: [Presence]
      summary: Renew the authenticated actor's HTTP device presence lease.
      description: >-
        Sets a bounded lease for this device independently of Ting delivery. Omitted or null
        activity clears the device's activity. A lease never grants message access and is not
        a delivery or read receipt. All requests revalidate IAM authority.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, data]
              properties:
                type: {type: string, const: renew_presence}
                data:
                  type: object
                  additionalProperties: false
                  properties:
                    activity:
                      type: [string, 'null']
                      enum: [typing, recording_voice, transcribing_voice, uploading_file, searching_gifs, null]
            example: {type: renew_presence, data: {activity: typing}}
      responses:
        '200':
          description: Renewed device deadline and aggregate actor presence.
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {type: string, const: renew_presence}
                  data: {$ref: '#/components/schemas/PresenceLease'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
    delete:
      operationId: closeDevicePresence
      tags: [Presence]
      summary: Close the authenticated actor's named HTTP presence lease.
      description: Does not close other devices or change Ting destinations. Repeated closure is harmless.
      responses:
        '204': {description: The device is no longer online through this lease.}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
...
  /api/v1/contracts:
    get:
      tags: [Operations]
      operationId: discoverContracts
      summary: Discover contract lifecycle and compatibility
      security: []
      parameters:
        - {$ref: '#/components/parameters/TestingKey'}
      responses:
        '200':
          description: Contract versions and compatibility matrix for the selected plane
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {type: string, const: contracts}
                  data:
                    type: object
                    required: [service, service_version, contracts, compatibility, delivery, retired_routes]
                    properties:
                      service: {type: string, const: silicon-dm}
                      service_version: {type: string}
                      contracts: {type: array, items: {type: object, additionalProperties: true}}
                      compatibility:
                        type: array
                        items:
                          type: object
                          required: [http]
                          properties: {http: {type: integer, const: 3}}
                      delivery: {$ref: '#/components/schemas/DeliveryDiscovery'}
                      retired_routes: {type: array, items: {type: string}}
                    additionalProperties: true
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/iam:
    get:
      tags: [Auth]
      operationId: iamInformation
      summary: Discover the public IAM application configuration
      security: []
      description: Available before login. Returns only public identifiers and URLs, never application secrets.
      parameters:
        - {$ref: '#/components/parameters/TestingKey'}
      responses:
        '200':
          description: Public application configuration
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [type, data]
                properties:
                  type: {type: string, const: iam}
                  data:
                    type: object
                    required: [app_id, iam_base_url, api_base_url, delivery]
                    properties:
                      app_id: {type: string}
                      iam_base_url: {type: string, format: uri}
                      api_base_url: {type: string, format: uri}
                      delivery: {$ref: '#/components/schemas/DeliveryDiscovery'}
                      testing_environment_id:
                        type: [string, 'null']
                        format: uuid
                      testing_generation:
                        type: [integer, 'null']
                      testing_environment:
                        type: [object, 'null']
                        additionalProperties: true
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/conversations/{conversation_id}/messages/{message_id}/receipts:
    post:
      tags: [Messages]
      operationId: recordMessageReceipt
      summary: Record a monotonic device receipt
      parameters:
        - {$ref: '#/components/parameters/OrgId'}
        - {$ref: '#/components/parameters/ConversationId'}
        - name: message_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/MessageReference'}
        - {$ref: '#/components/parameters/TestingKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, data]
              properties:
                type: {type: string, const: receipt}
                data:
                  type: object
                  required: [status, device_id]
                  properties:
                    status:
                      type: string
                      enum: [delivered, read]
                    device_id: {type: string, minLength: 1, maxLength: 255, pattern: '^[^\u0000-\u001F\u007F]{1,255}$'}
      responses:
        '200':
          description: Aggregate receipt state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [type, data]
                properties:
                  type: {type: string, const: receipt}
                  data: {$ref: '#/components/schemas/Message'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    honeycombService: {type: http, scheme: bearer, description: 'Dedicated Honeycomb service credential, independent of test sessions.'}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: IAM application access token (oat_)}
    testingEnvironmentKey:
      type: apiKey
      in: header
      name: X-Testing-Environment-Key
      description: 32-character DM testing root key. Selects an isolated data plane; normal actions still require its IAM user session.
  parameters:
    TestingGeneration:
      name: X-Testing-Environment-Generation
      in: header
      required: false
      schema: {type: integer, minimum: 1}
      description: Required for writes in a Honeycomb testing environment. Obtain from /api/v1/iam; stale generations fail with 409.
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema: {type: string}
    ConversationId:
      name: conversation_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/ConversationAddress'}
...
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    TestingKey:
      name: X-Testing-Environment-Key
      in: header
      required: false
      schema: {type: string, pattern: '^(ask_[A-Za-z0-9_-]{43}|[A-Za-z0-9]{32})$'}
      description: IAM test app secret for automatic discovery, or a legacy DM root key. Omit for production. Never send the IAM environment root key.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [type, data]
            properties:
              type: {type: string, const: error}
              data: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [type, data]
            properties:
              type: {type: string, const: error}
              data: {$ref: '#/components/schemas/Error'}
  schemas:
    DeliveryDiscovery:
      type: object
      required: [transport, app_id, api_base_url, browser_origin, receiver_authentication, publisher_authority, registration_path, sync_path, presence_path, receipts_path, dm_websocket_supported]
      properties:
        transport: {type: string, const: ting}
        app_id: {type: string, const: 'tos>ting'}
        api_base_url: {type: string, format: uri}
        browser_origin: {type: string, format: uri, const: 'https://ting.teamofsilicons.com'}
        receiver_authentication: {type: string, const: ting_session}
        publisher_authority: {type: string, const: originating_dm_session}
        registration_path: {type: string, const: /api/v1/delivery/registration}
        sync_path: {type: string, const: /api/v1/sync}
        presence_path: {type: string, const: '/api/v1/presence/devices/{device_id}'}
        receipts_path: {type: string, const: '/api/v1/conversations/{conversation_id}/messages/{message_id}/receipts'}
        dm_websocket_supported: {type: boolean, const: false}
...
    SyncEvent:
      type: object
      required: [event_id, sequence, type, conversation_id, message_id]
      properties:
        event_id: {type: string, format: uuid, description: Stable DM event identity; deduplicate independently of Ting IDs.}
        sequence: {type: integer, format: int64, minimum: 1}
        type: {type: string, enum: [message, message_status]}
        conversation_id: {$ref: '#/components/schemas/ConversationAddress'}
        message_id: {$ref: '#/components/schemas/MessageReference'}
    SyncPage:
      type: object
      required: [events, cursor, has_more, upper_sequence, testing_environment_id, testing_generation]
      properties:
        events: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/SyncEvent'}}
        cursor: {type: string, description: Opaque authenticated cursor; retain it even on an empty page.}
        has_more: {type: boolean}
        upper_sequence: {type: integer, format: int64, minimum: 0, description: Fixed upper boundary while paging this scan.}
        testing_environment_id: {type: [string, 'null'], format: uuid}
        testing_generation: {type: [integer, 'null'], minimum: 1}
    PresenceLease:
      type: object
      required: [presence, lease_expires_at, activity_expires_at]
      properties:
        presence: {$ref: '#/components/schemas/Presence'}
        lease_expires_at: {type: string, format: date-time}
        activity_expires_at: {type: [string, 'null'], format: date-time}
...
    ConversationAddress:
      type: string
      description: Immutable direct address, ordered Carbon before Silicon then lexically, or g:org:group-slug. UUID aliases remain accepted. Conversation
        routes also accept an authorized recipient account, resolving or creating the direct chat on send. Scope all addresses by organization and testing
        environment.
      examples: ['saket::cos:tos', 'g:tos:team']
...
    MemberRef:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [carbon, silicon]
        id:
          type: string
          description: IAM account ID.
          examples: [alice@example.com, 'assistant:acme']
...
    Message:
      type: object
      additionalProperties: false
      required: [message-id, conversation_id, recipient_id, sender, message, attachments, voice_transcript, reply, bundle, created_at, updated_at, deleted_at,
        delivered_at, read_at, history]
      properties:
        message-id: {$ref: '#/components/schemas/MessageCode'}
        conversation_id: {$ref: '#/components/schemas/ConversationAddress'}
        recipient_id: {type: string, description: 'History: the destination account or group. Callback: the receiving account, including the sender copy.'}
        sender: {$ref: '#/components/schemas/MemberRef'}
        message:
          type: [string, 'null']
          maxLength: 100000000
        attachments:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/AttachmentLink'}
        voice_transcript:
          type: [string, 'null']
          maxLength: 100000000
        reply:
          anyOf:
            - {$ref: '#/components/schemas/Reply'}
            - {type: 'null'}
        bundle:
          anyOf:
            - {$ref: '#/components/schemas/BundleRef'}
            - {type: 'null'}
        created_at: {type: string, format: date-time}
        updated_at:
          type: [string, 'null']
          format: date-time
        deleted_at:
          type: [string, 'null']
          format: date-time
        delivered_at:
          type: [string, 'null']
          format: date-time
        read_at:
          type: [string, 'null']
          format: date-time
        history:
          type: array
          items: {$ref: '#/components/schemas/MessageHistoryEntry'}
          description: Previous content, oldest first. Empty initially; deletion adds no history entry. Deleted responses hide content history.
      description: Fixed current snapshot. All fields are present. Text is empty for attachment-only sends, null for tombstones. Upsert by conversation_id,
        message-id and updated_at. Delivery records may hydrate the current edited or deleted snapshot on replay.
    BundleRef:
      type: object
      required: [id, role]
      properties:
        id: {$ref: '#/components/schemas/BundleCode'}
        role:
          type: string
          enum: [display, member]
...
    Presence:
      type: object
      required: [member_id, availability]
      properties:
        member_id: {type: string}
        availability:
          type: string
          enum: [online, offline]
        activity:
          type: [string, 'null']
          enum: [typing, recording_voice, transcribing_voice, uploading_file, searching_gifs, null]
        last_seen_at:
          type: [string, 'null']
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...
    MessageCode:
      type: string
      pattern: ^(?:[0-9a-z]{3}|[1-9a-z][0-9a-z]{3,12})$
      examples: ['000', 00a, zzz, '1000']
      description: Lowercase base36 message code, scoped to the conversation. Starts at 000, grows to 1000 after zzz; never reused.
    MessageReference:
      oneOf:
        - {$ref: '#/components/schemas/MessageCode'}
        - {type: string, format: uuid}
    AttachmentLink:
      type: string
      format: uri
      maxLength: 8192
      pattern: ^https://(?![^/?#]*@)[^/?#]+(?:[/?#]|$)
      description: A credential-free HTTPS attachment link. DM stores the URL without fetching it.
    Reply:
      type: object
      additionalProperties: false
      required: [message-id, sender, content]
      properties:
        message-id: {$ref: '#/components/schemas/MessageCode'}
        sender: {$ref: '#/components/schemas/MemberRef'}
        content:
          anyOf:
            - type: object
              required: [message, attachments, voice_transcript]
              properties:
                message:
                  type: [string, 'null']
                  maxLength: 100000000
                attachments:
                  type: array
                  maxItems: 100
                  items: {$ref: '#/components/schemas/AttachmentLink'}
                voice_transcript:
                  type: [string, 'null']
                  maxLength: 100000000
            - {type: 'null'}
          description: Server-resolved current reply content, or null after the referenced message is deleted.
...
    BundleCode:
      type: string
      pattern: ^(?:00[1-9a-z]|0[1-9a-z][0-9a-z]|[1-9a-z][0-9a-z]{2,})$
      description: Conversation-local sequential lowercase base36 bundle code. Starts at 001; increments for each new bundle and grows beyond three characters
        after zzz.
      examples: ['001', '002', 009, 00a, 00z, '010', zzz, '1000']
    DeviceId:
      type: string
      minLength: 1
      maxLength: 255
      pattern: ^[^\u0000-\u001F\u007F-\u009F]+$
      x-max-utf8-bytes: 255
      description: Client-chosen stable device/consumer ID. Reuse on reconnect and receipts; independent client installations should use distinct IDs. Accepts
        1–255 UTF-8 bytes, including spaces and Unicode, but no control characters. Not a hardware ID or credential.
      examples: [laptop-browser-1, phone-app-1, f96100cd-d86e-470b-88eb-4721830e59e1]
    MessageHistoryEntry:
      type: object
      additionalProperties: false
      required: [message, attachments, voice_transcript, reply, created_at]
      properties:
        message:
          type: [string, 'null']
          description: Text; empty for attachment-only messages; null after deletion.
        attachments:
          type: array
          items: {type: string, format: uri, pattern: '^https://'}
          description: File/audio/GIF URLs; empty when absent or deleted.
        voice_transcript:
          type: [string, 'null']
          description: Audio transcript; null when absent or deleted.
        reply:
          anyOf:
            - {$ref: '#/components/schemas/Reply'}
            - {type: 'null'}
        created_at: {type: string, format: date-time, description: When this content became current.}